code: working on ACL commands and also buildin the memory store a bit by bit

+++ src/storage/memory.rs
@@ -6,10 +6,13 @@ use std::{
 use log::warn;
 
 use crate::resp::value::Value;
+use super::entities::Entities;
 
 #[derive(Clone)]
 pub struct MemoryStore {
-  db: Arc<Mutex<HashMap<String, Value>>>,
+  general_store: Arc<Mutex<HashMap<String, Value>>>,
+  acl_store: Arc<Mutex<HashMap<String, HashMap<String, Entities>>>>,
+  authenticated: Arc<Mutex<bool>>,
 }
 
 pub trait Store {
@@ -22,35 +25,62 @@ pub trait Store {
 
 impl Store for MemoryStore {
   fn new() -> Self {
+    // @INFO Initialize the memory store
+    let acl_store = Arc::new(Mutex::new(HashMap::new()));
+    let general_store = Arc::new(Mutex::new(HashMap::new()));
+    let authenticated = Arc::new(Mutex::new(false));
+
+    // @INFO put the predefined entities inside the stores
+
     Self {
-      db: Arc::new(Mutex::new(HashMap::new())),
+      general_store,
+      acl_store,
+      authenticated,
     }
   }
 
   async fn set(&self, key: &str, value: Value) {
-    let mut db = self.db.lock().unwrap();
-    if db.contains_key(key) {
-      warn!("Key {} already exists, overwriting value", key);
-      db.insert(key.to_string(), value);
+    if self.authenticated.lock().unwrap().eq(&false) {
+      warn!("User not authenticated, using shared storage");
+      let mut db = self.general_store.lock().unwrap();
+      if db.contains_key(key) {
+        warn!("Key {} already exists, overwriting value", key);
+        db.insert(key.to_string(), value);
+      } else {
+        db.insert(key.to_string(), value);
+      }
     } else {
-      db.insert(key.to_string(), value);
+      // @TODO handle acl storage
+      todo!()
     }
   }
 
   async fn get(&self, key: &str) -> Option<Value> {
-    let db = self.db.lock().unwrap();
-    match db.get(key) {
-      Some(value) => Some(value.clone()),
-      None => Some(Value::Null),
+    if self.authenticated.lock().unwrap().eq(&false) {
+      warn!("User not authenticated, using shared storage");
+      let db = self.general_store.lock().unwrap();
+      match db.get(key) {
+        Some(value) => Some(value.clone()),
+        None => Some(Value::Null),
+      }
+    } else {
+      // @TODO handle acl storage
+      todo!()
     }
   }
 
   async fn delete(&self, key: &str) -> Option<Value> {
-    let mut db = self.db.lock().unwrap();
-    let result = db.remove(key);
-    if result.is_none() {
-      warn!("Key {} not found, cannot delete", key);
+    if self.authenticated.lock().unwrap().eq(&false) {
+      warn!("User not authenticated, using shared storage");
+      let mut db = self.general_store.lock().unwrap();
+      let result = db.remove(key);
+      if result.is_none() {
+        warn!("Key {} not found, cannot delete", key);
+      }
+      result
+    } else {
+      // @TODO handle acl storage
+      todo!()
     }
-    result
   }
 }
